refactor: move commons into models folder
