--- OnContractChange/error.ts
@@ -1,10 +1,0 @@
-// eslint-disable-next-line max-classes-per-file
-export class ValidationError extends Error {}
-export class FetchMembershipError extends Error {}
-export class FiscalCodeNotFoundError extends Error {}
-export class UpsertError extends Error {}
-export class FetchPecDelegatesError extends Error {}
-export class FetchPecAttachmentError extends Error {}
-export class FetchPecEmailError extends Error {}
-export class SaveContractError extends Error {}
-export class NotImplementedError extends Error {}
