+++ ProcessMembership/index.ts
@@ -1,6 +1,6 @@
 import { cosmosdb } from "../utils/cosmosdb";
 
-import { dao } from "../OnContractChange/dao";
+import { dao } from "../models/dao";
 import createHandler from "./handler";
 
 const handleContractChange = createHandler({ dao: dao(cosmosdb) });
